Refactor: Move all classes, methods into one file
