+++ src/ansys/aedt/core/extensions/hfss/mcad_assembly.py
@@ -23,11 +23,316 @@
 
 from pathlib import Path
 import tempfile
-import tkinter
 
-from ansys.aedt.core.extensions.hfss.src_mcad_assembly.frontend import MCADAssemblyFrontend
 from ansys.aedt.core.extensions.misc import ExtensionCommon
 from ansys.aedt.core.extensions.misc import get_arguments
+
+import os
+import tkinter
+from tkinter import ttk, filedialog
+import json
+
+import ansys.aedt.core
+from ansys.aedt.core.extensions.misc import ExtensionHFSSCommon
+from ansys.aedt.core.extensions.misc import get_aedt_version
+from ansys.aedt.core.extensions.misc import get_port
+from ansys.aedt.core.extensions.misc import get_process_id
+from ansys.aedt.core.extensions.misc import is_student
+from typing import Dict, List, Optional, Union
+
+from pydantic import BaseModel, Field
+
+from ansys.aedt.core.generic.constants import Axis
+
+DATA = {
+    "coordinate_system": {
+        "GLOBAL_2": {"origin": ["100mm", "0mm", "0mm"], "reference_cs": "Global"},
+        "CS_CLAMP": {"origin": ["-130mm", "80mm", "12mm"], "reference_cs": "GLOBAL_2"},
+    },
+    "assembly": {
+        "case": {
+            "component_type": "mcad",
+            "file_path": "E:\\models\\Chassi.a3dcomp",
+            "reference_coordinate_system": "Global",
+            "target_coordinate_system": "GLOBAL_2",
+            "arranges": [
+                {"operation": "rotate", "axis": "X", "angle": "0deg"},
+                {"operation": "move", "vector": ["0mm", "0mm", "0mm"]},
+            ],
+            "sub_components": {
+                "pcb": {
+                    "component_type": "ecad",
+                    "file_path": "E:\\models\\DCDC-Converter-App_main.aedbcomp",
+                    "target_coordinate_system": "Guiding_Pin",
+                    "layout_coordinate_systems": ["CABLE1_via_65", "CABLE2_via_65", "H0_via_65"],
+                    "reference_coordinate_system": "H0_via_65",
+                    "arranges": [
+                        {"operation": "rotate", "axis": "X", "angle": "0deg"},
+                        {"operation": "move", "vector": ["0mm", "0mm", "0mm"]},
+                    ],
+                    "sub_components": {
+                        "cable_1": {
+                            "component_type": "mcad",
+                            "file_path": "E:\\models\\Cable_1.a3dcomp",
+                            "target_coordinate_system": "CABLE1_via_65",
+                        },
+                        "cable_2": {
+                            "component_type": "mcad",
+                            "file_path": "E:\\models\\Cable_1.a3dcomp",
+                            "target_coordinate_system": "CABLE2_via_65",
+                        },
+                    },
+                }
+            },
+        },
+        "clamp_monitor": {
+            "component_type": "mcad",
+            "file_path": "E:\\models\\BCI_MONITORING_CLAMP.a3dcomp",
+            "reference_coordinate_system": "Global",
+            "target_coordinate_system": "CS_CLAMP",
+        },
+    },
+}
+
+# Frontend
+class AedtInfo(BaseModel):
+    version: str = ""
+    port: int
+    aedt_process_id: Union[int, None]
+    student_version: bool = False
+
+class MCADAssemblyFrontend(ExtensionHFSSCommon):
+    EXTENSION_TITLE = "MCAD Assembly"
+    GRID_PARAMS = {"padx": 15, "pady": 10, "sticky": "nsew"}
+    PACK_PARAMS = {"padx": 15, "pady": 10}
+
+    tab_frame_main = None
+
+    config_data: dict = dict()
+
+    def __init__(self, withdraw: bool = False):
+        self.aedt_info = AedtInfo(
+            port=get_port(), version=get_aedt_version(), aedt_process_id=get_process_id(), student_version=is_student()
+        )
+
+        super().__init__(
+            self.EXTENSION_TITLE,
+            theme_color="light",
+            withdraw=withdraw,
+            add_custom_content=True,
+            toggle_row=2,
+            toggle_column=0,
+        )
+
+    def add_toggle_theme_button(self, parent, toggle_row, toggle_column):
+        return
+
+    def add_toggle_theme_button_(self, parent):
+        """Create a button to toggle between light and dark themes."""
+        button_frame = ttk.Frame(
+            parent, style="PyAEDT.TFrame", relief=tkinter.SUNKEN, borderwidth=2, name="theme_button_frame"
+        )
+        button_frame.pack(fill="both", expand=False, **{"padx": 5, "pady": 5})
+
+        ttk.Button(
+            button_frame,
+            width=10,
+            text="Run",
+            command=lambda: self.run(self.config_data),
+            style="PyAEDT.TButton",
+            name="run",
+        ).pack(anchor="w", side="left", **{"padx": 15, "pady": 10})
+
+        self._widgets["button_frame"] = button_frame
+
+        change_theme_button = ttk.Button(
+            button_frame,
+            width=10,
+            text="\u263d",
+            command=self.toggle_theme,
+            style="PyAEDT.TButton",
+            name="theme_toggle_button",
+        )
+        # change_theme_button.grid(row=0, column=0, **{"padx": 15, "pady": 10})
+        change_theme_button.pack(anchor="e", side="right", **{"padx": 15, "pady": 10})
+        self._widgets["change_theme_button"] = change_theme_button
+
+    def add_extension_content(self):
+        self.root.geometry("700x600")
+
+        menubar = tkinter.Menu(self.root)
+        self.root.config(menu=menubar)
+
+        nb = ttk.Notebook(self.root, name="notebook", style="PyAEDT.TNotebook")
+        self.tab_frame_main = ttk.Frame(nb, name="main", style="PyAEDT.TFrame")
+
+        nb.add(self.tab_frame_main, text="Main")
+
+        nb.pack(fill="both", expand=True)
+
+        create_tab_main(self.tab_frame_main, self)
+        self.add_toggle_theme_button_(self.root)
+
+    def run(self, config_data):
+        hfss = ansys.aedt.core.Hfss(**self.aedt_info.model_dump())
+        app = MCADAssemblyBackend.load(data=config_data)
+        app.run(hfss)
+        del app
+
+        if "PYTEST_CURRENT_TEST" not in os.environ:  # pragma: no cover
+            hfss.release_desktop(False, False)
+        else:
+            hfss.close_project(save=False)
+        return
+
+
+# create main tab
+def create_tab_main(tab_frame, master):
+    tree = ttk.Treeview(tab_frame, name="tree")
+    tree.pack(expand=True, fill="both", **master.PACK_PARAMS)
+
+    ttk.Button(
+        tab_frame,
+        # width=10,
+        text="Load Configure File",
+        command=lambda: load_dict(tree, master),
+        style="PyAEDT.TButton",
+        name="load",
+    ).pack(anchor="w", **master.PACK_PARAMS)
+
+
+def load_dict(tree, master):
+    file_path = filedialog.askopenfilename(
+        title="Select Design",
+        filetypes=(("JSON", "*.json"), ("All files", "*.*")),
+    )
+    if not file_path:  # pragma: no cover
+        return
+    else:
+        with open(file_path, "r") as f:
+            master.config_data = json.load(f)
+    tree.delete(*tree.get_children())  # clear everything
+    insert_items(tree, "", master.config_data)  # insert new data
+
+
+def insert_items(tree, parent, dictionary):
+    if isinstance(dictionary, dict):
+        for key, value in dictionary.items():
+            node = tree.insert(parent, "end", text=str(key), open=False)
+            insert_items(tree, node, value)
+    elif isinstance(dictionary, list):
+        for idx, item in enumerate(dictionary):
+            node = tree.insert(parent, "end", text=f"[{idx}]", open=False)
+            insert_items(tree, node, item)
+    else:
+        tree.insert(parent, "end", text=str(dictionary))
+
+# end of create_tab_main function
+
+# End of frontend
+
+# Below is the backend for the MCAD assembly extension.
+class Arrange(BaseModel):
+    operation: str
+    axis: Optional[str] = "X"
+    angle: Optional[str] = "0deg"
+    vector: Optional[list[Union[str]]] = ["0mm", "0mm", "0mm"]
+
+    class Config:
+        extra = "forbid"
+
+
+class Component(BaseModel):
+    component_type: str
+    name: str
+    file_path: str
+    reference_coordinate_system: str = "Global"
+    target_coordinate_system: str
+    layout_coordinate_systems: Optional[List[str]] = Field(default_factory=list)
+    arranges: List[Arrange] = Field(default_factory=list)
+    sub_components: Optional[Dict] = Field(default_factory=dict)
+
+    class Config:
+        extra = "forbid"
+
+    @classmethod
+    def load(cls, name, data):
+        sub_components = {name: cls.load(name, comp) for name, comp in data.get("sub_components", {}).items()}
+        data_ = data.copy()
+        data_["sub_components"] = sub_components
+        data_["name"] = name
+        return cls(**data_)
+
+    def assemble_sub_components(self, hfss, cs_prefix=""):
+        for name, comp in self.sub_components.items():
+            comp.assemble(hfss, cs_prefix)
+
+    def apply_arrange(self, hfss):
+        for i in self.arranges:
+            if i.operation == "rotate":
+                hfss.modeler.rotate(self.name, getattr(Axis, i.axis), i.angle)
+            elif i.operation == "move":
+                hfss.modeler.move(self.name, i.vector)
+
+    def assemble(self, hfss, cs_prefix=None):
+        """
+        Parameters
+        ----------
+         cs_prefix : str
+            This is the name of the component definition.
+        """
+        if cs_prefix:
+            self.target_coordinate_system = f"{cs_prefix}_{self.target_coordinate_system}"
+
+        if self.component_type == "mcad":
+            comp = hfss.modeler.insert_3d_component(
+                name=self.name,
+                input_file=self.file_path,
+                coordinate_system=self.target_coordinate_system,
+            )
+            temp = None
+        else:
+            comp = hfss.modeler.insert_layout_component(
+                input_file=self.file_path,
+                coordinate_system=self.target_coordinate_system,
+                reference_coordinate_system=self.reference_coordinate_system,
+                layout_coordinate_systems=self.layout_coordinate_systems,
+                name=self.name,
+            )
+            temp = comp.definition_name
+
+        if comp is False:
+            raise ValueError(self.name, self.file_path, self.target_coordinate_system)
+
+        self.apply_arrange(hfss)
+        if self.sub_components:
+            self.assemble_sub_components(hfss, cs_prefix=temp)
+
+
+Component.model_rebuild()
+
+
+class MCADAssemblyBackend(BaseModel):
+    coordinate_system: Dict[str, Dict[str, Union[str, List[str]]]] = Field(default_factory=dict)
+    sub_components: Dict[str, Component] = Field(default_factory=dict)
+
+    class Config:
+        extra = "forbid"
+
+    @classmethod
+    def load(cls, data):
+        return cls(
+            coordinate_system=data.get("coordinate_system", {}),
+            sub_components={name: Component.load(name, comp) for name, comp in data.get("assembly", {}).items()},
+        )
+
+    def run(self, hfss):
+        for name, value in self.coordinate_system.items():
+            hfss.modeler.create_coordinate_system(name=name, **value)
+
+        for name, comp in self.sub_components.items():
+            comp.assemble(hfss)
+# End of MCADAssemblyBackend
 
 if __name__ == "__main__":  # pragma: no cover
     args = get_arguments()
