--- src/ansys/aedt/core/extensions/hfss/src_mcad_assembly/tab_main.py
@@ -1,68 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# Copyright (C) 2021 - 2025 ANSYS, Inc. and/or its affiliates.
-# SPDX-License-Identifier: MIT
-#
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-import json
-from tkinter import filedialog
-from tkinter import ttk
-
-
-def create_tab_main(tab_frame, master):
-    tree = ttk.Treeview(tab_frame, name="tree")
-    tree.pack(expand=True, fill="both", **master.PACK_PARAMS)
-
-    ttk.Button(
-        tab_frame,
-        # width=10,
-        text="Load Configure File",
-        command=lambda: load_dict(tree, master),
-        style="PyAEDT.TButton",
-        name="load",
-    ).pack(anchor="w", **master.PACK_PARAMS)
-
-
-def load_dict(tree, master):
-    file_path = filedialog.askopenfilename(
-        title="Select Design",
-        filetypes=(("JSON", "*.json"), ("All files", "*.*")),
-    )
-    if not file_path:  # pragma: no cover
-        return
-    else:
-        with open(file_path, "r") as f:
-            master.config_data = json.load(f)
-    tree.delete(*tree.get_children())  # clear everything
-    insert_items(tree, "", master.config_data)  # insert new data
-
-
-def insert_items(tree, parent, dictionary):
-    if isinstance(dictionary, dict):
-        for key, value in dictionary.items():
-            node = tree.insert(parent, "end", text=str(key), open=False)
-            insert_items(tree, node, value)
-    elif isinstance(dictionary, list):
-        for idx, item in enumerate(dictionary):
-            node = tree.insert(parent, "end", text=f"[{idx}]", open=False)
-            insert_items(tree, node, item)
-    else:
-        tree.insert(parent, "end", text=str(dictionary))
