+++ tests/system/extensions/test_mcad_assembly.py
@@ -27,8 +27,8 @@ from pathlib import Path
 import pytest
 
 from ansys.aedt.core import Hfss
-from ansys.aedt.core.extensions.hfss.src_mcad_assembly.backend import MCADAssemblyBackend
-from ansys.aedt.core.extensions.hfss.src_mcad_assembly.template import DATA
+from ansys.aedt.core.extensions.hfss.mcad_assembly import MCADAssemblyBackend
+from ansys.aedt.core.extensions.hfss.mcad_assembly import DATA
 
 MODEL_FOLDER = Path(__file__).parent / "example_models" / "mcad_assembly"
